refactoring locust
